Removed rendering-related code from the NoteStem class, so that it can also be used when calculating the staff layout (currently, note stems are not included when computing the height of a staff)
git-svn-id: http://svn.powertabs.org/svn/ptedev/trunk@647 5b1460fe-822e-4aa1-ac1b-8a6412c92c8c

+++ source/painters/systemrenderer.h
@@ -117,7 +117,7 @@ private:
     void adjustAccidentals(QMultiMap<int, StdNotationPainter*>& accidentalsMap);
     void drawMultiBarRest(boost::shared_ptr<const Barline> currentBarline,
                           const StaffData& currentStaffInfo, int measureCount);
-    void drawLedgerLines(const std::vector<double> &noteLocations, const double xLocation,
+    void drawLedgerLines(const std::vector<int> &noteLocations, const double xLocation,
                          const StaffData& staffData);
 };
